[lookupform] Refactor all demo related names.
